Development (#3)
Adding in the following milestone elements:

  1.  Tmux resurrect capabilities.
  2.  Env injection support.

Tmux resurrect runs by default on the host machine when the application is launched. This can be disabled with the -disabled=true flag.

Env's can now be injected, per process, using the config.yaml file.

The contents of the .env will be loaded in per application that is denoted in the yaml config. If you want the same env for multiple apps, you will need to replicate the given line, inserting the .env into the app.

Tmux resurrect runs the saving code to store the session data 1x per minute. (If not disabled)

View the README.md for more precise instructions on running the given commands